+++ src/jupyterscad/_render.py
@@ -14,60 +14,17 @@ PARTICULAR PURPOSE. See the GNU General Public License for more details.
 You should have received a copy of the GNU General Public License along with
 this program. If not, see <https://www.gnu.org/licenses/>.
 """
+import logging
+import subprocess
 import tempfile
 from os import PathLike
+from pathlib import Path
+from shutil import which
 from typing import Optional, Union
 
-import pythreejs as pjs
+from .exceptions import OpenSCADError, RenderError
 
-from ._openscad import process
-from ._visualize import visualize_stl
-from .exceptions import RenderError
-
-
-def render(
-    obj,
-    width: int = 400,
-    height: int = 400,
-    grid_unit: float = 1,
-    outfile: Optional[Union[str, PathLike]] = None,
-    openscad_exec: Optional[Union[str, PathLike]] = None,
-) -> pjs.Renderer:
-    """Render a visusualization of an OpenSCAD object.
-
-    Typical usage example:
-
-        >>> display(render(cube(3)))
-
-    Args:
-        obj: OpenSCAD object to visualize.
-        width: Visualization pixel width on page.
-        height: Visualization pixel height on page.
-        grid_unit: Grid cell size, 0 to disable, -1 for automatic
-        outfile: Name of stl file to generate. No stl file is generated if None.
-        openscad_exec: Path to openscad executable.
-
-    Returns:
-        Rendering to be displayed.
-
-    Raises:
-        exceptions.OpenSCADException: An error occurred running OpenSCAD.
-    """
-    try:
-        if outfile:
-            render_stl(obj, outfile, openscad_exec=openscad_exec)
-            r = visualize_stl(outfile, width=width, height=height, grid_unit=grid_unit)
-        else:
-            with tempfile.NamedTemporaryFile(
-                suffix=".stl", delete=False
-            ) as stl_tmp_file:
-                render_stl(obj, stl_tmp_file.name, openscad_exec=openscad_exec)
-                r = visualize_stl(
-                    stl_tmp_file.name, width=width, height=height, grid_unit=grid_unit
-                )
-        return r
-    except RenderError as e:
-        e.show()
+LOGGER = logging.getLogger(__name__)
 
 
 def render_stl(
@@ -94,3 +51,42 @@ def render_stl(
             fp.write(str(obj))
 
         process(scad_tmp_file.name, outfile, executable=openscad_exec)
+
+
+def process(scad_file, output_file, executable: Optional[Union[str, PathLike]] = None):
+    """Generate stl from scad using OpenSCAD executable"""
+    if executable:
+        executable = Path(executable)
+        if not executable.is_file():
+            raise OpenSCADError(f"Specified executable {executable} does not exist.")
+    else:
+        executable = detect_executable()
+
+    cmd = [executable, "-o", output_file, scad_file]
+    LOGGER.info(cmd)
+    try:
+        out = subprocess.run(cmd, check=True, capture_output=True, text=True)
+        if "ERROR" in out.stderr:
+            with open(scad_file) as fp:
+                scad_str = fp.read()
+            raise RenderError(message=out.stderr, src=scad_str)
+    except subprocess.CalledProcessError as e:
+        raise OpenSCADError(str(e.stderr))
+
+
+def detect_executable() -> Path:
+    """Detect the OpenSCAD executable"""
+
+    detected_executable = which("openscad") or which(
+        "openscad", path="/Applications/OpenSCAD.app/Contents/MacOS"
+    )  # macOS
+
+    if not detected_executable:
+        raise OpenSCADError(
+            "OpenSCAD executable autodetection failed. "
+            "Please specify the path to the OpenSCAD executable."
+        )
+
+    LOGGER.debug(f"Executable path ({detected_executable}) found.")
+
+    return Path(detected_executable)